Add files via upload
On 8 November 2016, the Government of India announced the demonetisation of all ₹500 and ₹1000 banknotes of the Mahatma Gandhi Series. The government claimed that the action would curtail the shadow economy and crack down on the use of illicit and counterfeit cash to fund illegal activity and terrorism. The sudden nature of the announcement and the prolonged cash shortages in the weeks that followed created significant disruption throughout the economy, threatening economic output, the effects of which can be seen even in current sentiment of the people 

This report is based on the analysis of RBI’s official transactional data across various payment methods over three years of time (2015-2018) and also the twitter data for tweets of past week as well as for the short span of time in the past when demonetisation was announced.

The objectives that are targeted in this report are as follows:
•	The sentimental analysis of people towards demonetisation which is compared across two different time zones i.e., for the current and for the past period when demonetization was declared also sentiments over digital transactions as a whole.
•	The impact on socio economic factors by demonetisation and the trend they have followed over the years.
•	The various payment methods that are available for and how there volume (in millions) and value (in billions ) has affected over the years.
•	The impact of one payment method over the year in terms of volume and if at all one has contributed to the other over the year.


The results are targeted both granularly on demonetisation and also summarising digital transactions as a whole. Python is used for textual analysis part and R is used for numeric data manipulation. 


The future scope lies in predicting  the volume and value of various payment methods and also the predicting various socio economic trends that are payment indicators might follow.